2to3 -f long
This is a funny one! All occurances of float are in an isinstance() call that includes int. I'll fix those in a follow-up commit.

+++ tests/rtmplite/multitask.py
@@ -321,7 +321,7 @@ class FDReady(YieldCondition):
 
 
 def _is_file_descriptor(fd):
-    return isinstance(fd, (int, long))
+    return isinstance(fd, (int, int))
 
 
 def readable(fd, timeout=None):
